+++ relay-general/src/protocol/session.rs
@@ -1,4 +1,4 @@
-use std::fmt;
+use std::fmt::{self, Display};
 use std::time::SystemTime;
 
 use chrono::{DateTime, Utc};
@@ -6,11 +6,11 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::macros::derive_fromstr_and_display;
+use crate::protocol::utils::null_to_default;
 use crate::protocol::IpAddr;
 
 /// The type of session event we're dealing with.
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
 pub enum SessionStatus {
     /// The session is healthy.
     ///
@@ -24,6 +24,8 @@ pub enum SessionStatus {
     Abnormal,
     /// The session exited cleanly but experienced some errors during its run.
     Errored,
+    /// Unknown status, for forward compatibility.
+    Unknown(String),
 }
 
 impl SessionStatus {
@@ -40,6 +42,39 @@ impl SessionStatus {
     /// Returns `true` if the status indicates a fatal session.
     pub fn is_fatal(&self) -> bool {
         matches!(self, SessionStatus::Crashed | SessionStatus::Abnormal)
+    }
+    fn as_str(&self) -> &str {
+        match self {
+            SessionStatus::Ok => "ok",
+            SessionStatus::Crashed => "crashed",
+            SessionStatus::Abnormal => "abnormal",
+            SessionStatus::Exited => "exited",
+            SessionStatus::Errored => "errored",
+            SessionStatus::Unknown(s) => s.as_str(),
+        }
+    }
+}
+
+impl Display for SessionStatus {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
+
+relay_common::impl_str_serde!(SessionStatus, "A session status");
+
+impl std::str::FromStr for SessionStatus {
+    type Err = ParseSessionStatusError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s {
+            "ok" => SessionStatus::Ok,
+            "crashed" => SessionStatus::Crashed,
+            "abnormal" => SessionStatus::Abnormal,
+            "exited" => SessionStatus::Exited,
+            "errored" => SessionStatus::Errored,
+            other => SessionStatus::Unknown(other.to_owned()),
+        })
     }
 }
 
@@ -60,14 +95,6 @@ impl fmt::Display for ParseSessionStatusError {
 }
 
 impl std::error::Error for ParseSessionStatusError {}
-
-derive_fromstr_and_display!(SessionStatus, ParseSessionStatusError, {
-    SessionStatus::Ok => "ok",
-    SessionStatus::Crashed => "crashed",
-    SessionStatus::Abnormal => "abnormal",
-    SessionStatus::Exited => "exited",
-    SessionStatus::Errored => "errored",
-});
 
 #[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "snake_case")]
@@ -153,7 +180,6 @@ pub trait SessionLike {
     fn abnormal_count(&self) -> u32;
     fn crashed_count(&self) -> u32;
     fn all_errors(&self) -> Option<SessionErrored>;
-    fn final_duration(&self) -> Option<(f64, SessionStatus)>;
     fn abnormal_mechanism(&self) -> AbnormalMechanism;
 }
 
@@ -189,7 +215,11 @@ pub struct SessionUpdate {
     #[serde(rename = "attrs")]
     pub attributes: SessionAttributes,
     /// The abnormal mechanism.
-    #[serde(default, skip_serializing_if = "AbnormalMechanism::is_none")]
+    #[serde(
+        default,
+        deserialize_with = "null_to_default",
+        skip_serializing_if = "AbnormalMechanism::is_none"
+    )]
     pub abnormal_mechanism: AbnormalMechanism,
 }
 
@@ -230,15 +260,6 @@ impl SessionLike for SessionUpdate {
             SessionStatus::Crashed => 1,
             _ => 0,
         }
-    }
-
-    fn final_duration(&self) -> Option<(f64, SessionStatus)> {
-        if self.status.is_terminal() {
-            if let Some(duration) = self.duration {
-                return Some((duration, self.status));
-            }
-        }
-        None
     }
 
     fn all_errors(&self) -> Option<SessionErrored> {
@@ -301,10 +322,6 @@ impl SessionLike for SessionAggregateItem {
         self.crashed
     }
 
-    fn final_duration(&self) -> Option<(f64, SessionStatus)> {
-        None
-    }
-
     fn all_errors(&self) -> Option<SessionErrored> {
         // Errors contain crashed & abnormal as well.
         // See https://github.com/getsentry/snuba/blob/c45f2a8636f9ea3dfada4e2d0ae5efef6c6248de/snuba/migrations/snuba_migrations/sessions/0003_sessions_matview.py#L80-L81
@@ -344,9 +361,22 @@ impl SessionAggregates {
 
 #[cfg(test)]
 mod tests {
+
+    use std::str::FromStr;
+
     use similar_asserts::assert_eq;
 
     use super::*;
+
+    #[test]
+    fn test_sessionstatus_unknown() {
+        let unknown = SessionStatus::from_str("invalid status").unwrap();
+        if let SessionStatus::Unknown(inner) = unknown {
+            assert_eq!(inner, "invalid status".to_owned());
+        } else {
+            panic!();
+        }
+    }
 
     #[test]
     fn test_session_default_values() {
@@ -494,6 +524,23 @@ mod tests {
   "started": "2020-02-07T14:16:00Z",
   "status": "abnormal",
   "abnormal_mechanism": "invalid_mechanism",
+  "attrs": {
+    "release": "sentry-test@1.0.0",
+    "environment": "production"
+  }
+}"#;
+
+        let update = SessionUpdate::parse(json.as_bytes()).unwrap();
+        assert_eq!(update.abnormal_mechanism, AbnormalMechanism::None);
+    }
+
+    #[test]
+    fn test_session_null_abnormal_mechanism() {
+        let json = r#"{
+  "sid": "8333339f-5675-4f89-a9a0-1c935255ab58",
+  "started": "2020-02-07T14:16:00Z",
+  "status": "abnormal",
+  "abnormal_mechanism": null,
   "attrs": {
     "release": "sentry-test@1.0.0",
     "environment": "production"